Implement random walk environment
The random walk environment can be completed very fast and it also works
nicely to test agents that estimate a baseline. To allow this this
random walk environment has been slightly modified from the one from the
literature. In this case instead of terminating on both ends, it
terminates on the right goal and after a fixed length of steps which is
twice the length of the walk. This way an agent can ramp up more
negative reward more easily, which can be nicely counteracted by
baseline estimates.