Add alternation environment
This is a simple environment, which is easy to solve. It provides a
quick means of sanity checking learning algorithms making sure they
learn anything. The simplicity of the environment makes it easier to
understand any obvious problems an algorithm might have. Additionally,
it is quick to execute. A random agent will always perform poorly
compared to an agent successfully learning the one winning strategy of
this environment.