Provide reward dependent on knob distance
With the previous reward system the agent would have no indication if it
did better or worse. By providing the negative L2 norm of action and
knob vector as penalty agents can better gauge if they are getting
closer to a solution. Before the reward space was too uniform and the
chance that an agent would observe any reward improvement was very slim,
hence agents couldn't properly learn the environment.